--- packages/openswap/scripts/flowgen.sh
@@ -1,3 +1,0 @@
-for i in $(find dist -type f -name "*.d.ts");
-  do sh -c "npx flowgen $i -o ${i%.*.*}.js.flow";
-done;
