+++ packages/swap/src/adapters/openswap-api/api.ts
@@ -1,128 +1,128 @@
-import {Swap} from '@yoroi/types'
-import {OpenSwap, OpenSwapApi} from '@yoroi/openswap'
+import {AxiosInstance} from 'axios'
+import {
+  cancelOrder, // returns an unsigned transaction to cancel the order.
+  createOrder, // returns a datum and a contract address to create the order transaction.
+  getCompletedOrders,
+  getOrders, // returns all orders for a given stake key hash.
+} from './orders'
+import {getTokenPairs} from './token-pairs'
+import {getTokens} from './tokens'
+import {
+  CancelOrderRequest,
+  CreateOrderRequest,
+  Network,
+  Provider,
+  PriceAddress,
+  TokenAddress,
+} from './types'
+import {axiosClient} from './config'
+import {getPrice} from './price'
+import {getLiquidityPools, getPoolsPair} from './pools'
 
-import {transformersMaker} from '../../helpers/transformers'
-import {apiMocks} from './api.mocks'
-
-export const swapApiMaker = (
-  {
-    isMainnet,
-    stakingKey,
-    primaryTokenId,
-    supportedProviders,
-  }: {
-    isMainnet?: boolean
-    stakingKey: string
-    primaryTokenId: string
-    supportedProviders: ReadonlyArray<Swap.SupportedProvider>
-  },
-  deps?: {openswap?: OpenSwapApi},
-): Readonly<Swap.Api> => {
-  const api =
-    deps?.openswap ?? new OpenSwapApi(isMainnet ? 'mainnet' : 'preprod')
-  const transformers = transformersMaker(primaryTokenId)
-
-  const getOpenOrders: Swap.Api['getOpenOrders'] = () =>
-    api
-      .getOrders(stakingKey)
-      .then((orders) =>
-        orders.map((order) => transformers.asYoroiOpenOrder(order)),
+export class OpenSwapApi {
+  constructor(
+    public readonly network: Network,
+    private readonly client: AxiosInstance = axiosClient,
+  ) {
+    if (!supportedNetworks.includes(network)) {
+      throw new Error(
+        `Supported networks are ${supportedNetworks.join(
+          ', ',
+        )}, got ${network}`,
       )
-
-  const getCompletedOrders: Swap.Api['getCompletedOrders'] = () =>
-    api
-      .getCompletedOrders(stakingKey)
-      .then((orders) =>
-        orders.map((order) => transformers.asYoroiCompletedOrder(order)),
-      )
-
-  const createOrder: Swap.Api['createOrder'] = async (orderData) => {
-    const {amounts, address, selectedPool} = orderData
-
-    const orderRequest: OpenSwap.CreateOrderRequest = {
-      walletAddress: address,
-      protocol:
-        selectedPool.provider as OpenSwap.CreateOrderRequest['protocol'],
-      poolId: selectedPool.poolId,
-      sell: transformers.asOpenswapAmount(amounts.sell),
-      buy: transformers.asOpenswapAmount(amounts.buy),
     }
-
-    return api.createOrder(orderRequest).then((response) => {
-      if (response.status === 'failed')
-        return Promise.reject(response.reason ?? 'Unknown error')
-
-      return {
-        datum: response.datum,
-        datumHash: response.hash,
-        contractAddress: response.address,
-      }
-    })
   }
 
-  const cancelOrder: Swap.Api['cancelOrder'] = (orderData) =>
-    api
-      .cancelOrder({
-        orderUTxO: orderData.utxos.order,
-        collateralUTxO: orderData.utxos.collateral,
-        walletAddress: orderData.address,
-      })
-      .then((response) => response)
-
-  const getTokenPairs: Swap.Api['getTokenPairs'] = async (token) =>
-    !isMainnet
-      ? apiMocks.getTokenPairs // preprod doesn't return any tokens
-      : api
-          .getTokenPairs(transformers.asOpenswapTokenId(token))
-          .then(transformers.asYoroiBalanceTokens)
-
-  const getTokens: Swap.Api['getTokens'] = async () => {
-    return api.getTokens().then(transformers.asYoroiBalanceTokenInfos)
+  public async createOrder(orderData: CreateOrderRequest) {
+    return createOrder({network: this.network, client: this.client}, orderData)
   }
 
-  const getPools: Swap.Api['getPools'] = async ({
+  public async cancelOrder(orderData: CancelOrderRequest) {
+    return cancelOrder({network: this.network, client: this.client}, orderData)
+  }
+
+  public async getOrders(stakeKeyHash: string) {
+    return getOrders(
+      {network: this.network, client: this.client},
+      {stakeKeyHash},
+    )
+  }
+
+  public async getCompletedOrders(stakeKeyHash: string) {
+    return getCompletedOrders(
+      {network: this.network, client: this.client},
+      {stakeKeyHash},
+    )
+  }
+
+  public async getPrice({
+    baseToken,
+    quoteToken,
+  }: {
+    baseToken: PriceAddress
+    quoteToken: PriceAddress
+  }) {
+    return getPrice(
+      {network: this.network, client: this.client},
+      {baseToken, quoteToken},
+    )
+  }
+
+  public async getPoolsPair({
     tokenA,
     tokenB,
-    providers = supportedProviders,
-  }) => {
-    if (!isMainnet) return apiMocks.getPools // preprod doesn't return any pools
-
-    return api
-      .getLiquidityPools({
-        tokenA,
-        tokenB,
-        providers,
-      })
-      .then(transformers.asYoroiPools)
+  }: {
+    tokenA: TokenAddress
+    tokenB: TokenAddress
+  }) {
+    return getPoolsPair(
+      {network: this.network, client: this.client},
+      {tokenA, tokenB},
+    )
   }
 
-  const getPrice: Swap.Api['getPrice'] = async ({baseToken, quoteToken}) => {
-    const opBaseToken = transformers.asOpenswapPriceTokenAddress(baseToken)
-    const opQuoteToken = transformers.asOpenswapPriceTokenAddress(quoteToken)
-
-    return api
-      .getPrice({
-        baseToken: {
-          ...opBaseToken,
-        },
-        quoteToken: {
-          ...opQuoteToken,
-        },
-      })
-      .then((response) => response.price)
+  public async getLiquidityPools({
+    tokenA,
+    tokenB,
+    providers,
+  }: {
+    tokenA: string
+    tokenB: string
+    providers: ReadonlyArray<Provider>
+  }) {
+    return getLiquidityPools(
+      {network: this.network, client: this.client},
+      {tokenA, tokenB, providers},
+    )
   }
 
-  return {
-    getPrice,
-    getOpenOrders,
-    cancelOrder,
-    createOrder,
-    getTokens,
-    getTokenPairs,
-    getPools,
-    getCompletedOrders,
-    stakingKey,
-    primaryTokenId,
-    supportedProviders,
-  } as const
+  public async getTokenPairs({policyId = '', assetName = ''} = {}) {
+    const tokenPairs = await getTokenPairs(
+      {network: this.network, client: this.client},
+      {policyId, assetName},
+    )
+
+    return tokenPairs
+  }
+
+  public async getTokens() {
+    return getTokens({network: this.network, client: this.client})
+  }
 }
+
+export const supportedNetworks: ReadonlyArray<Network> = [
+  'mainnet',
+  'preprod',
+] as const
+
+export const supportedProviders: ReadonlyArray<Provider> = [
+  'minswap',
+  'muesliswap_v1',
+  'muesliswap_v3',
+  'muesliswap_v4',
+  'spectrum',
+  'sundaeswap',
+  'vyfi',
+  'wingriders',
+  'muesliswap_v2',
+] as const
