+++ yoroi.code-workspace
@@ -37,10 +37,6 @@
       "path": "./packages/common"
     },
     {
-      "name": "@yoroi/openswap",
-      "path": "./packages/openswap"
-    },
-    {
       "name": "@yoroi/staking",
       "path": "./packages/staking"
     },
